change(web): compatibility analyzer -> own file
